tests: update tests
