+++ test/react.test.js
@@ -3,7 +3,7 @@
  */
 
 import renderer from "react-test-renderer";
-import {ThemeProvider, useTheme, styled} from "@siimple/react";
+import {ThemeProvider, useTheme, styled, Box, classNames} from "@siimple/react";
 
 const selector = `style[data-siimple="siimple-react"]`;
 
@@ -12,6 +12,18 @@ describe("ThemeProvider", () => {
         const component = renderer.create((
             <ThemeProvider theme={{}}>
                 <div>Hello world</div>
+            </ThemeProvider>
+        ));
+
+        expect(component.toJSON()).toMatchSnapshot();
+    });
+});
+
+describe("Box", () => {
+    it("should render", () => {
+        const component = renderer.create((
+            <ThemeProvider theme={{}}>
+                <Box css={{color: "white"}}>Hello world</Box>
             </ThemeProvider>
         ));
 
@@ -73,5 +85,26 @@ describe("styled", () => {
         expect(div).toBeDefined();
         expect(div.props.className).toEqual(expect.stringContaining("sii-"));
         expect(document.querySelector(selector).innerHTML).toEqual(expect.stringContaining("color:black"));
+    });
+});
+
+describe("classNames", () => {
+    it("should join arguments into a single string", () => {
+        const className = classNames("foo", "bar", "baz");
+        expect(className).toBe("foo bar baz");
+    });
+
+    it("should add only classes with a truthly value in an object", () => {
+        const className = classNames({
+            foo: true,
+            bar: false,
+            baz: true,
+        });
+        expect(className).toBe("foo baz");
+    });
+
+    it("should join all elements in an array", () => {
+        const className = classNames(["foo", "bar"], "baz");
+        expect(className).toBe("foo bar baz");
     });
 });
